2022/5/2  22:10 二维码支付调试成功，目前剩下付款吗 还有表格无法操作功能、背景图片功能，历史销售记录导出功能

+++ teashop_project/src/main/java/com/guet/controller/AddProduct.java
@@ -67,8 +67,8 @@ public class AddProduct extends JFrame {
         }else {
             Tea tea = new Tea();
             tea.setTeaName(addTextField1.getText());
-            tea.setTeaDiscount(Double.parseDouble(addTextField2.getText()));
-            tea.setTeaPrice(Integer.parseInt(addTextField3.getText()));
+            tea.setTeaDiscount(Float.parseFloat(addTextField2.getText()));
+            tea.setTeaPrice(Float.parseFloat(addTextField3.getText()));
             tea.setTeaCategory(addTextField4.getText());
             tea.setTeaAmount(Integer.parseInt(addTextField5.getText()));
             System.out.println(tea);
