+++ teashop_project/src/main/java/com/guet/dao/Impl/DiscountProductDaoImpl.java
@@ -32,9 +32,9 @@ public class DiscountProductDaoImpl implements DiscountProductDao {
                 tea.setTeaId(rs.getInt("tea_id"));
                 tea.setTeaName(rs.getString("tea_name"));
                 tea.setTeaAmount(rs.getInt("tea_amount"));
-                tea.setTeaPrice(rs.getInt("tea_price"));
+                tea.setTeaPrice(rs.getFloat("tea_price"));
                 tea.setTeaCategory(rs.getString("tea_category"));
-                tea.setTeaDiscount(rs.getDouble("tea_discount"));
+                tea.setTeaDiscount(rs.getFloat("tea_discount"));
                 list.add(tea);
             }
         } catch (Exception e) {
@@ -74,7 +74,7 @@ public class DiscountProductDaoImpl implements DiscountProductDao {
             while(rs.next()){
                 Tea tea=new Tea();
                 tea.setTeaName(rs.getString("tea_name"));
-                tea.setTeaDiscount(rs.getDouble("tea_discount"));
+                tea.setTeaDiscount(rs.getFloat("tea_discount"));
                 list.add(tea);
             }
             System.out.println(list);
