+++ teashop_project/src/main/java/com/guet/dao/Impl/ProductDaoImpl.java
@@ -1,8 +1,9 @@
 package com.guet.dao.Impl;
 
+
+
 import com.guet.dao.ProductDao;
 import com.guet.entity.Tea;
-import com.guet.util.ConnUtil;
 import com.guet.util.ConnectionHandler;
 
 import java.sql.Connection;
@@ -36,9 +37,9 @@ public class ProductDaoImpl implements ProductDao {
                 tea.setTeaId(rs.getInt("tea_id"));
                 tea.setTeaName(rs.getString("tea_name"));
                 tea.setTeaAmount(rs.getInt("tea_amount"));
-                tea.setTeaPrice(rs.getDouble("tea_price"));
+                tea.setTeaPrice(rs.getFloat("tea_price"));
                 tea.setTeaCategory(rs.getString("tea_category"));
-                tea.setTeaDiscount(rs.getDouble("tea_discount"));
+                tea.setTeaDiscount(rs.getFloat("tea_discount"));
                 list.add(tea);
             }
         } catch (Exception e) {
@@ -67,9 +68,9 @@ public class ProductDaoImpl implements ProductDao {
                 tea.setTeaId(rs.getInt("tea_id"));
                 tea.setTeaName(rs.getString("tea_name"));
                 tea.setTeaAmount(rs.getInt("tea_amount"));
-                tea.setTeaPrice(rs.getInt("tea_price"));
+                tea.setTeaPrice(rs.getFloat("tea_price"));
                 tea.setTeaCategory(rs.getString("tea_category"));
-                tea.setTeaDiscount(rs.getDouble("tea_discount"));
+                tea.setTeaDiscount(rs.getFloat("tea_discount"));
                 list.add(tea);
             }
         } catch (Exception e) {
